fix: Run command init as late as possible
Due to differences in GATT timing on Android versus desktop, running
init in the middle of the service/characteristic bringup sequence
caused a race between GATT functions, meaning we'd error out of the
chain completely and nothing worked. Adding init to the end and
waiting on it to resolve means we'll have a clean connection when we
pass the object off.

Fixes #2